requirements.txt
PyOpenGL
PyOpenGL_accelerate
numpy>=1.14.2
c3d>=0.3.0
matplotlib>=2.0.2
pygame>=1.9.3
scipy>=0.19.1
noise>=1.2.2
Pillow>=5.0.0
PySignal==1.1.1
pytest>=3.4.2
scikit_learn>=0.18.2
transformations
